checkpont: utils.py로 분리
