+++ business/views.py
@@ -35,61 +35,6 @@ from google import genai
 from django.conf import settings
 
 
-def extract_chat_title(path: str) -> str:
-    """
-    텍스트 파일 path의 첫 줄에서
-    “~님과” 앞부분만 가져옵니다.
-    """
-    with open(path, "r", encoding="utf-8") as f:
-        first_line = (
-            f.readline().strip()
-        )  # ex: "🦁멋사 13기 잡담방🦁 님과 카카오톡 대화"
-
-    # '(.*?)' : 가능한 한 짧게 매칭, '님과' 앞까지 캡쳐
-    match = re.match(r"^(.*?)\s*님과", first_line)
-    if match:
-        return match.group(1)
-    else:
-        # “님과” 패턴이 없으면 줄 전체를 리턴하거나 빈 문자열
-        return first_line
-
-def count_chat_participants_with_gemini(file_path: str) -> int:
-    """
-    Gemini API를 사용해 채팅 로그 파일의 참여 인원 수를 계산합니다.
-    - file_path: 분석할 채팅 파일의 절대 경로
-    - 반환값: 계산된 인원 수 (정수)
-    """
-    try:
-        # 파일이 매우 클 경우를 대비해 앞부분 일부만 읽는 것이 효율적입니다.
-        with open(file_path, "r", encoding="utf-8") as f:
-            # 여기서는 최대 500줄만 읽도록 제한 (성능 및 비용 최적화)
-            lines = f.readlines()
-            chat_content_sample = "".join(lines[:500])
-
-        # Gemini API 클라이언트 초기화
-        client = genai.Client(api_key=settings.GEMINI_API_KEY)
-        response = client.models.generate_content(
-            model='gemini-2.0-flash',
-            contents=[
-                "당신은 카카오톡 채팅 로그 분석 전문가입니다. \
-                주어진 채팅 내용에서 고유한 참여자(사람 이름)가 총 몇 명인지 세어주세요. \
-                아래 채팅 내용을 보고, 다른 부가적인 설명은 일절 하지 말고, 오직 최종 인원 수를 나타내는 정수 숫자만 답변해주세요."]
-            + [chat_content_sample]
-        )
-
-        # Gemini의 응답(e.g., "15" 또는 "총 15명")에서 숫자만 추출하여 정수로 변환
-        numbers = re.findall(r'\d+', response.text)
-        if numbers:
-            return int(numbers[0])
-        else:
-            # 숫자를 찾지 못한 경우 기본값 반환
-            return 1
-
-    except Exception as e:
-        # API 호출 실패, 응답 파싱 실패 등 예외 발생 시
-        print(f"Gemini로 인원 수 분석 중 에러 발생: {e}")
-        # 기본값 혹은 에러 처리에 맞는 값을 반환합니다. 여기서는 1을 반환.
-        return 1
 
 # Create your views here.
 class BusChatView(APIView):
